Add files via upload
This is all the JavaScript code and assets for my interactive planets project done with the Myo gesture control band. One library has been deprecated since I originally built it and as such the project no longer works properly, I have managed to fix part of it but am not in a position to properly fix it all.